Open the dashboard by clicking the toast
