+++ src/lib/tabs.js
@@ -6,3 +6,16 @@ export function getCurrentTab() {
     .query({ active: true, currentWindow: true })
     .then(tabs => tabs[0]);
 }
+
+export function openDashboardTab() {
+  let dashboard_url = browser.runtime.getURL('dashboard/index.html');
+  browser.tabs.query({ currentWindow: true, url: dashboard_url }).then(e => {
+    if (e.length < 1) {
+      browser.tabs.create({
+        url: dashboard_url,
+      });
+    } else {
+      browser.tabs.update(e[0].id, { active: true });
+    }
+  });
+}
